feat(collections): authors, refactoring
