+++ styles/pages/collection.module.scss
@@ -2,12 +2,17 @@
 @import 'styles/variables';
 
 .collection-page {
-  text-decoration: none; // placeholder.
+  padding-bottom: 3rem;
+  padding-top: 3rem;
 
   &__heading {
     display: flex;
     flex-direction: row;
     padding-bottom: 46px;
+
+    @include screen-mobile {
+      flex-direction: column;
+    }
   }
 
   &__cover {
@@ -15,10 +20,15 @@
     height: 90px;
     margin-right: 29px;
     width: 90px;
+
+    @include screen-mobile {
+      margin-bottom: 30px;
+      margin-right: 0;
+    }
   }
 
   &__text {
-    text-decoration: none;
+    text-decoration: none; // keeper.
   }
 
   &__subheading {
@@ -35,12 +45,12 @@
     line-height: 60px;
 
     @include screen-tablet-touch-mobile {
-      font-size: 24px;
-      line-height: 30px;
+      font-size: 30px;
+      line-height: 36px;
     }
   }
 
-  &__author {
+  &__authors {
     color: $primary;
     font-size: 16px;
     line-height: 24px;
