+++ styles/pages/tag.module.scss
@@ -5,16 +5,6 @@
   padding-bottom: 3rem;
   padding-top: 3rem;
 
-  &__level-a,
-  &__level-b,
-  &__level-c,
-  &__level-d {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    padding-bottom: 2rem;
-  }
-
   &__topic {
     font-size: $default-font-size;
     line-height: 24px;
@@ -29,34 +19,7 @@
   }
 }
 
-@mixin screens-below-desktop {
-  .tag-page {
-    &__level-b,
-    &__level-c {
-      padding-bottom: 0;
-    }
-
-    &__card {
-      padding-bottom: 25px;
-    }
-  }
-}
-
-@include screen-tablet-large {
-  @include screens-below-desktop;
-}
-
-@include screen-tablet {
-  @include screens-below-desktop;
-}
-
-@include screen-touch {
-  @include screens-below-desktop;
-}
-
 @include screen-mobile {
-  @include screens-below-desktop;
-
   .tag-page {
     &__title {
       font-size: 36px;
